requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
PyYAML==6.0.2
